Add another integration test
